app/requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
django==4.0
gunicorn==20.1.0
djangorestframework==3.12.4
python-decouple==3.4
asgiref==3.5.0
sqlparse==0.4.2
psycopg2-binary==2.9.1
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2021.1
pyzmq==22.1.0
